Add brute force algorithm
BruteForce.c contains an algorithm that finds a solution to the clock by testing
every possible permutation.

cmdclock.c is a new library that contains methods to create a clock from command
line arguments. main.c was updated to use this library. BruteForce.c depends on
this as well

Makefile was updated to build the brute force program